refactor: move whatsapp group handler into platform
